NFDIV-3147: split callbacks to new event
