+++ src/integrationTest/java/uk/gov/hmcts/divorce/solicitor/SystemSolicitorPaymentIT.java
@@ -1,7 +1,6 @@
 package uk.gov.hmcts.divorce.solicitor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import feign.FeignException;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -15,8 +14,8 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import uk.gov.hmcts.divorce.common.config.WebMvcConfig;
+import uk.gov.hmcts.divorce.notification.NotificationService;
 import uk.gov.hmcts.divorce.payment.model.CreditAccountPaymentResponse;
-import uk.gov.hmcts.divorce.systemupdate.service.CcdUpdateService;
 import uk.gov.hmcts.divorce.testutil.CaseDataWireMock;
 import uk.gov.hmcts.divorce.testutil.FeesWireMock;
 import uk.gov.hmcts.divorce.testutil.IdamWireMock;
@@ -25,11 +24,11 @@ import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 
 import java.io.IOException;
 
-import static java.util.Objects.requireNonNull;
 import static net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson;
 import static net.javacrumbs.jsonunit.assertj.JsonAssertions.json;
 import static net.javacrumbs.jsonunit.core.Option.IGNORING_EXTRA_FIELDS;
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -38,23 +37,15 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static uk.gov.hmcts.divorce.divorcecase.model.SolicitorPaymentMethod.FEE_PAY_BY_ACCOUNT;
 import static uk.gov.hmcts.divorce.divorcecase.model.State.Draft;
 import static uk.gov.hmcts.divorce.payment.model.PaymentStatus.SUCCESS;
-import static uk.gov.hmcts.divorce.solicitor.event.SolicitorSubmitApplication.SOLICITOR_SUBMIT;
-import static uk.gov.hmcts.divorce.testutil.FeesWireMock.stubForFeesNotFound;
-import static uk.gov.hmcts.divorce.testutil.IdamWireMock.SYSTEM_USER_ROLE;
-import static uk.gov.hmcts.divorce.testutil.IdamWireMock.stubForIdamDetails;
-import static uk.gov.hmcts.divorce.testutil.IdamWireMock.stubForIdamToken;
+import static uk.gov.hmcts.divorce.systemupdate.event.SystemSolicitorPayment.SYSTEM_SOLICITOR_PAYMENT;
 import static uk.gov.hmcts.divorce.testutil.PaymentWireMock.stubCreditAccountPayment;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_START_URL;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.AUTHORIZATION;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.AUTH_HEADER_VALUE;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.SERVICE_AUTHORIZATION;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.SUBMITTED_URL;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.SYSTEM_USER_USER_ID;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SERVICE_AUTH_TOKEN;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SYSTEM_AUTHORISATION_TOKEN;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.callbackRequest;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseDataWithOrderSummary;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseDataWithStatementOfTruth;
@@ -71,7 +62,7 @@ import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.expectedResponse;
     IdamWireMock.PropertiesInitializer.class,
     CaseDataWireMock.PropertiesInitializer.class,
     PaymentWireMock.PropertiesInitializer.class})
-public class SolicitorSubmitApplicationIT {
+public class SystemSolicitorPaymentIT {
 
     @Autowired
     private MockMvc mockMvc;
@@ -86,7 +77,7 @@ public class SolicitorSubmitApplicationIT {
     private AuthTokenGenerator serviceTokenGenerator;
 
     @MockBean
-    private CcdUpdateService ccdUpdateService;
+    private NotificationService notificationService;
 
     @BeforeAll
     static void setUp() {
@@ -102,29 +93,6 @@ public class SolicitorSubmitApplicationIT {
         CaseDataWireMock.stopAndReset();
         FeesWireMock.stopAndReset();
         PaymentWireMock.stopAndReset();
-    }
-
-    @Test
-    public void givenFeeEventIsNotAvailableWhenCallbackIsInvokedThenReturn404FeeEventNotFound()
-        throws Exception {
-        stubForFeesNotFound();
-
-        mockMvc.perform(post(ABOUT_TO_START_URL)
-                .contentType(APPLICATION_JSON)
-                .header(SERVICE_AUTHORIZATION, AUTH_HEADER_VALUE)
-                .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-                .content(objectMapper.writeValueAsString(callbackRequest(caseDataWithOrderSummary(), SOLICITOR_SUBMIT)))
-                .accept(APPLICATION_JSON))
-            .andExpect(
-                status().isNotFound()
-            )
-            .andExpect(
-                result -> assertThat(result.getResolvedException()).isExactlyInstanceOf(FeignException.NotFound.class)
-            )
-            .andExpect(
-                result -> assertThat(requireNonNull(result.getResolvedException()).getMessage())
-                    .contains("404 Fee event not found")
-            );
     }
 
     @Test
@@ -153,7 +121,7 @@ public class SolicitorSubmitApplicationIT {
                 .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
                 .content(objectMapper.writeValueAsString(callbackRequest(
                     data,
-                    SOLICITOR_SUBMIT,
+                    SYSTEM_SOLICITOR_PAYMENT,
                     Draft.name())))
                 .accept(APPLICATION_JSON))
             .andExpect(
@@ -164,6 +132,8 @@ public class SolicitorSubmitApplicationIT {
         assertThatJson(mvcResult.getResponse().getContentAsString())
             .when(IGNORING_EXTRA_FIELDS)
             .isEqualTo(json(expectedCcdAboutToSubmitCallbackResponse()));
+
+        verifyNoMoreInteractions(notificationService);
     }
 
     @Test
@@ -179,7 +149,7 @@ public class SolicitorSubmitApplicationIT {
                 .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
                 .content(objectMapper.writeValueAsString(callbackRequest(
                     caseData,
-                    SOLICITOR_SUBMIT,
+                    SYSTEM_SOLICITOR_PAYMENT,
                     Draft.name())))
                 .accept(APPLICATION_JSON))
             .andExpect(
@@ -188,51 +158,12 @@ public class SolicitorSubmitApplicationIT {
             .andExpect(
                 content().json(expectedCcdAboutToSubmitCallbackErrorResponse())
             );
-    }
 
-    @Test
-    void givenValidCaseDataWhenSubmittedCallbackIsInvokedThenCallSyaytemPaymentEvent()
-        throws Exception {
-
-        stubCreditAccountPayment(
-            CREATED,
-            CreditAccountPaymentResponse
-                .builder()
-                .status(SUCCESS.toString())
-                .caseReference(TEST_CASE_ID.toString())
-                .build()
-        );
-
-        var data = caseDataWithStatementOfTruth();
-        data.getApplication().setApplicationPayments(null);
-        data.getApplication().setSolPaymentHowToPay(FEE_PAY_BY_ACCOUNT);
-        data.getApplication().setPbaNumbers(getPbaNumbersForAccount("PBA0012345"));
-        data.getApplication().setApplicationFeeOrderSummary(orderSummaryWithFee());
-        data.getApplicant1().getSolicitor().setOrganisationPolicy(organisationPolicy());
-
-        stubForIdamDetails(TEST_SYSTEM_AUTHORISATION_TOKEN, SYSTEM_USER_USER_ID, SYSTEM_USER_ROLE);
-        stubForIdamToken(TEST_SYSTEM_AUTHORISATION_TOKEN);
-
-        MvcResult mvcResult = mockMvc.perform(post(SUBMITTED_URL)
-                .contentType(APPLICATION_JSON)
-                .header(SERVICE_AUTHORIZATION, TEST_SERVICE_AUTH_TOKEN)
-                .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-                .content(objectMapper.writeValueAsString(callbackRequest(
-                    data,
-                    SOLICITOR_SUBMIT,
-                    Draft.name())))
-                .accept(APPLICATION_JSON))
-            .andExpect(
-                status().isOk()
-            )
-            .andReturn();
-
-        assertThatJson(mvcResult.getResponse().getContentAsString())
-            .isEqualTo("{}");
+        verifyNoInteractions(notificationService);
     }
 
     private String expectedCcdAboutToSubmitCallbackResponse() throws IOException {
-        return expectedResponse("classpath:wiremock/responses/about-to-submit-statement-of-truth.json");
+        return expectedResponse("classpath:wiremock/responses/about-to-submit-statement-of-truth-payment.json");
     }
 
     private String expectedCcdAboutToSubmitCallbackErrorResponse() throws IOException {
